Refactor API.
