+++ src/archive.rs
@@ -1,403 +1,25 @@
 //! An archive of bufkit soundings.
 
-use std::{
-    io::{Read, Write},
-    path::{Path, PathBuf},
-    str::FromStr,
-};
-
-use crate::{
-    errors::BufkitDataErr,
-    inventory::Inventory,
-    models::Model,
-    site::{Site, StateProv},
-};
+#[cfg(test)]
+use crate::{BufkitDataErr, Inventory, Model, Site, StateProv};
+use std::path::PathBuf;
 
 /// The archive.
 #[derive(Debug)]
 pub struct Archive {
     root: PathBuf,                 // The root directory.
-    data_root: PathBuf,            // the directory containing the downloaded files.
     db_conn: rusqlite::Connection, // An sqlite connection.
 }
 
+mod add_data;
 mod clean;
+mod query;
+mod root;
 
 impl Archive {
-    const DATA_DIR: &'static str = "data";
-    const DB_FILE: &'static str = "index.sqlite";
-
-    /// Initialize a new archive.
-    pub fn create(root: &dyn AsRef<Path>) -> Result<Self, BufkitDataErr> {
-        let data_root = root.as_ref().join(Archive::DATA_DIR);
-        let db_file = root.as_ref().join(Archive::DB_FILE);
-        let root = root.as_ref().to_path_buf();
-
-        std::fs::create_dir_all(&data_root)?; // The folder to store the sounding files.
-
-        // Create and set up the archive
-        let db_conn = rusqlite::Connection::open_with_flags(
-            db_file,
-            rusqlite::OpenFlags::SQLITE_OPEN_READ_WRITE | rusqlite::OpenFlags::SQLITE_OPEN_CREATE,
-        )?;
-
-        db_conn.execute_batch(include_str!("create_index.sql"))?;
-
-        Ok(Archive {
-            root,
-            data_root,
-            db_conn,
-        })
-    }
-
-    /// Open an existing archive.
-    pub fn connect(root: &dyn AsRef<Path>) -> Result<Self, BufkitDataErr> {
-        let data_root = root.as_ref().join(Archive::DATA_DIR);
-        let db_file = root.as_ref().join(Archive::DB_FILE);
-        let root = root.as_ref().to_path_buf();
-
-        // Create and set up the archive
-        let db_conn = rusqlite::Connection::open_with_flags(
-            db_file,
-            rusqlite::OpenFlags::SQLITE_OPEN_READ_WRITE,
-        )?;
-
-        Ok(Archive {
-            root,
-            data_root,
-            db_conn,
-        })
-    }
-
-    /// Retrieve a path to the root. Allows caller to store files in the archive.
-    pub fn root(&self) -> &Path {
-        &self.root
-    }
-
-    /// Retrieve a list of sites in the archive.
-    pub fn sites(&self) -> Result<Vec<Site>, BufkitDataErr> {
-        let mut stmt = self.db_conn.prepare(
-            "
-                SELECT 
-                    sites.station_num,
-                    site_ids.id,
-                    name,
-                    state,
-                    notes,
-                    auto_download,
-                    tz_offset_sec 
-            FROM sites JOIN site_ids ON sites.station_num = site_ids.station_num",
-        )?;
-
-        let vals: Result<Vec<Site>, BufkitDataErr> = stmt
-            .query_and_then(rusqlite::NO_PARAMS, Self::parse_row_to_site)?
-            .map(|res| res.map_err(BufkitDataErr::Database))
-            .collect();
-
-        vals
-    }
-
-    /// Retrieve the information about a single site id
-    pub fn site_info(&self, station_num: u32) -> Option<Site> {
-        self.db_conn
-            .query_row_and_then(
-                "
-                    SELECT 
-                         sites.station_num,
-                         site_ids.id,
-                         name,
-                         state,
-                         notes,
-                         auto_download,
-                         tz_offset_sec
-                    FROM sites LEFT JOIN site_ids ON sites.station_num = site_ids.station_num
-                    WHERE sites.station_num = ?1
-                ",
-                &[&station_num],
-                Self::parse_row_to_site,
-            )
-            .ok()
-    }
-
-    /// Given a site_id string, get the corresponding Site object.
-    pub fn site_for_id(&self, site_id: &str) -> Option<Site> {
-        self.db_conn
-            .query_row_and_then(
-                "
-                    SELECT 
-                         sites.station_num,
-                         site_ids.id,
-                         name,
-                         state,
-                         notes,
-                         auto_download,
-                         tz_offset_sec
-                    FROM site_ids JOIN sites ON site_ids.station_num = sites.station_num
-                    WHERE site_ids.id = ?1
-                ",
-                &[&site_id.to_uppercase()],
-                Self::parse_row_to_site,
-            )
-            .ok()
-    }
-
-    fn parse_row_to_site(row: &rusqlite::Row) -> Result<Site, rusqlite::Error> {
-        let station_num: u32 = row.get(0)?;
-        let id: Option<String> = row.get(1)?;
-        let name: Option<String> = row.get(2)?;
-        let notes: Option<String> = row.get(4)?;
-        let auto_download: bool = row.get(5)?;
-        let state: Option<StateProv> = row
-            .get::<_, String>(3)
-            .ok()
-            .and_then(|a_string| StateProv::from_str(&a_string).ok());
-
-        let time_zone: Option<chrono::FixedOffset> =
-            row.get::<_, i32>(6).ok().map(|offset: i32| {
-                if offset < 0 {
-                    chrono::FixedOffset::west(offset.abs())
-                } else {
-                    chrono::FixedOffset::east(offset)
-                }
-            });
-
-        Ok(Site {
-            station_num,
-            id,
-            name,
-            notes,
-            state,
-            auto_download,
-            time_zone,
-        })
-    }
-
-    /// Given a site, get the current ID used for that site.
-    pub fn id_for_site(&self, site: &Site) -> Option<String> {
-        self.db_conn
-            .query_row_and_then(
-                "SELECT id from site_ids WHERE station_num = ?1",
-                &[&site.station_num],
-                |row| row.get(0),
-            )
-            .ok()
-    }
-
-    /// Modify a site's values.
-    pub fn set_site_info(&self, site: &Site) -> Result<(), BufkitDataErr> {
-        self.db_conn.execute(
-            "
-                UPDATE sites
-                SET (state,name,notes,auto_download,tz_offset_sec)
-                = (?2, ?3, ?4, ?5, ?6)
-                WHERE station_num = ?1
-            ",
-            &[
-                &site.station_num,
-                &site.state.map(|state_prov| state_prov.as_static_str())
-                    as &dyn rusqlite::types::ToSql,
-                &site.name,
-                &site.notes,
-                &site.auto_download,
-                &site.time_zone.map(|tz| tz.local_minus_utc()),
-            ],
-        )?;
-
-        self.check_update_site_id(site)
-    }
-
-    /// Add a site to the list of sites.
-    pub fn add_site(&self, site: &Site) -> Result<(), BufkitDataErr> {
-        self.db_conn.execute(
-            "INSERT INTO sites (station_num, state, name, notes, auto_download, tz_offset_sec)
-                  VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-            &[
-                &site.station_num as &dyn rusqlite::ToSql,
-                &site.state.map(|state_prov| state_prov.as_static_str())
-                    as &dyn rusqlite::types::ToSql,
-                &site.name,
-                &site.notes,
-                &site.auto_download,
-                &site.time_zone.map(|tz| tz.local_minus_utc()),
-            ],
-        )?;
-
-        self.check_update_site_id(site)?;
-
-        Ok(())
-    }
-
-    #[inline]
-    fn check_update_site_id(&self, site: &Site) -> Result<(), BufkitDataErr> {
-        if let Some(ref site_id) = site.id {
-            let mut needs_insert = true;
-
-            if let Some(other_site) = self.site_for_id(site_id) {
-                if other_site.station_num != site.station_num {
-                    self.db_conn.execute(
-                        "DELETE FROM site_ids WHERE station_num = ?1",
-                        &[&other_site.station_num],
-                    )?;
-                } else {
-                    needs_insert = false;
-                }
-            }
-
-            if needs_insert {
-                self.db_conn.execute(
-                    "INSERT INTO site_ids (station_num, id) VALUES (?1, ?2)",
-                    &[
-                        &site.station_num as &dyn rusqlite::ToSql,
-                        &site_id.to_uppercase() as &dyn rusqlite::ToSql,
-                    ],
-                )?;
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Check if a site already exists
-    pub fn site_exists(&self, station_num: u32) -> Result<bool, BufkitDataErr> {
-        let number: i32 = self.db_conn.query_row(
-            "SELECT COUNT(*) FROM sites WHERE station_num = ?1",
-            &[&station_num],
-            |row| row.get(0),
-        )?;
-
-        Ok(number == 1)
-    }
-
-    /// Get a list of all the available model initialization times for a given site and model.
-    pub fn init_times(
-        &self,
-        site: &Site,
-        model: Model,
-    ) -> Result<Vec<chrono::NaiveDateTime>, BufkitDataErr> {
-        let mut stmt = self.db_conn.prepare(
-            "
-                SELECT init_time FROM files
-                WHERE station_num = ?1 AND model = ?2
-                ORDER BY init_time ASC
-            ",
-        )?;
-
-        let init_times: Vec<Result<chrono::NaiveDateTime, _>> = stmt
-            .query_map(
-                &[
-                    &site.station_num as &dyn rusqlite::ToSql,
-                    &model.as_static_str() as &dyn rusqlite::ToSql,
-                ],
-                |row| row.get::<_, chrono::NaiveDateTime>(0),
-            )?
-            .map(|res| res.map_err(BufkitDataErr::Database))
-            .collect();
-
-        let init_times: Vec<chrono::NaiveDateTime> =
-            init_times.into_iter().filter_map(Result::ok).collect();
-
-        Ok(init_times)
-    }
-
-    /// Get the number of values files in the archive for the model and intitialization time.
-    pub fn count_init_times(&self, site: &Site, model: Model) -> Result<i64, BufkitDataErr> {
-        let num_records: i64 = self.db_conn.query_row(
-            "
-                SELECT COUNT(init_time) FROM files
-                WHERE station_num = ?1 AND model = ?2
-            ",
-            &[
-                &site.station_num as &dyn rusqlite::ToSql,
-                &model.as_static_str() as &dyn rusqlite::ToSql,
-            ],
-            |row| row.get(0),
-        )?;
-
-        Ok(num_records)
-    }
-
-    /// Get an inventory of soundings for a site & model.
-    pub fn inventory(&self, site: &Site, model: Model) -> Result<Inventory, BufkitDataErr> {
-        let init_times = self.init_times(site, model)?;
-        Inventory::new(init_times, model, site)
-    }
-
-    /// Get a list of models in the archive for this site.
-    pub fn models(&self, site: &Site) -> Result<Vec<Model>, BufkitDataErr> {
-        let mut stmt = self
-            .db_conn
-            .prepare("SELECT DISTINCT model FROM files WHERE station_num = ?1")?;
-
-        let vals: Result<Vec<Model>, BufkitDataErr> = stmt
-            .query_map(&[&site.station_num], |row| row.get::<_, String>(0))?
-            .map(|res| res.map_err(BufkitDataErr::Database))
-            .map(|res| {
-                res.and_then(|name| Model::from_str(&name).map_err(BufkitDataErr::StrumError))
-            })
-            .collect();
-
-        vals
-    }
-
-    /// Retrieve the model initialization time of the most recent model in the archive.
-    pub fn most_recent_init_time(
-        &self,
-        site: &Site,
-        model: Model,
-    ) -> Result<chrono::NaiveDateTime, BufkitDataErr> {
-        let init_time: chrono::NaiveDateTime = self.db_conn.query_row(
-            "
-                SELECT init_time FROM files
-                WHERE station_num = ?1 AND model = ?2
-                ORDER BY init_time DESC
-                LIMIT 1
-            ",
-            &[
-                &site.station_num as &dyn rusqlite::ToSql,
-                &model.as_static_str() as &dyn rusqlite::ToSql,
-            ],
-            |row| row.get(0),
-        )?;
-
-        Ok(init_time)
-    }
-
-    /// Retrieve all the initialization times of all sounding files that have a sounding with a
-    /// valid time in the specified range (inclusive).
-    pub fn init_times_for_soundings_valid_between(
-        &self,
-        start: chrono::NaiveDateTime,
-        end: chrono::NaiveDateTime,
-        site: &Site,
-        model: Model,
-    ) -> Result<Vec<chrono::NaiveDateTime>, BufkitDataErr> {
-        let mut stmt = self.db_conn.prepare(
-            "
-                SELECT init_time
-                FROM files
-                WHERE station_num = ?1 AND model = ?2 AND init_time <= ?4 AND end_time >= ?3
-                ORDER BY init_time ASC
-            ",
-        )?;
-
-        let init_times: Result<Vec<chrono::NaiveDateTime>, _> = stmt
-            .query_map(
-                &[
-                    &site.station_num as &dyn rusqlite::types::ToSql,
-                    &model.as_static_str() as &dyn rusqlite::types::ToSql,
-                    &start as &dyn rusqlite::types::ToSql,
-                    &end as &dyn rusqlite::types::ToSql,
-                ],
-                |row| row.get::<_, chrono::NaiveDateTime>(0),
-            )?
-            .map(|res| res.map_err(BufkitDataErr::Database))
-            .collect();
-
-        init_times
-    }
-
     /// Check to see if a file is present in the archive and it is retrieveable.
-    pub fn file_exists(
+    #[cfg(test)]
+    fn file_exists(
         &self,
         site: &Site,
         model: Model,
@@ -415,201 +37,12 @@ impl Archive {
 
         Ok(num_records == 1)
     }
-
-    /// Get the number of files stored in the archive.
-    pub fn count(&self) -> Result<i64, BufkitDataErr> {
-        let num_records: i64 =
-            self.db_conn
-                .query_row("SELECT COUNT(*) FROM files", rusqlite::NO_PARAMS, |row| {
-                    row.get(0)
-                })?;
-
-        Ok(num_records)
-    }
-
-    /// Add a bufkit file to the archive.
-    pub fn add(
-        &self,
-        site: &Site,
-        model: Model,
-        init_time: chrono::NaiveDateTime,
-        end_time: chrono::NaiveDateTime,
-        text_data: &str,
-    ) -> Result<(), BufkitDataErr> {
-        let site_id = if let Some(ref site_id) = site.id {
-            site_id
-        } else {
-            return Err(BufkitDataErr::InvalidSiteId("None".to_owned()));
-        };
-
-        if !self.site_exists(site.station_num)? {
-            self.add_site(&Site {
-                station_num: site.station_num,
-                id: site.id.clone(),
-                name: None,
-                notes: None,
-                state: None,
-                auto_download: false,
-                time_zone: None,
-            })?;
-        }
-
-        if let Some(Site {
-            station_num: station_check,
-            ..
-        }) = self.site_for_id(site_id)
-        {
-            if station_check != site.station_num {
-                self.db_conn.execute(
-                    "DELETE FROM site_ids WHERE station_num = ?1",
-                    &[&station_check],
-                )?;
-                self.db_conn.execute(
-                    "INSERT INTO site_ids (station_num, id) VALUES (?1, ?2)",
-                    &[
-                        &site.station_num as &dyn rusqlite::ToSql,
-                        &site_id.to_uppercase() as &dyn rusqlite::ToSql,
-                    ],
-                )?;
-            }
-        }
-
-        let file_name = self.compressed_file_name(site_id, model, init_time);
-        let file = std::fs::File::create(self.data_root.join(&file_name))?;
-        let mut encoder = flate2::write::GzEncoder::new(file, flate2::Compression::default());
-        encoder.write_all(text_data.as_bytes())?;
-
-        self.db_conn.execute(
-            "INSERT OR REPLACE INTO files (station_num, model, init_time, end_time, file_name)
-                  VALUES (?1, ?2, ?3, ?4, ?5)",
-            &[
-                &site.station_num as &dyn rusqlite::types::ToSql,
-                &model.as_static_str() as &dyn rusqlite::types::ToSql,
-                &init_time as &dyn rusqlite::types::ToSql,
-                &end_time,
-                &file_name,
-            ],
-        )?;
-
-        Ok(())
-    }
-
-    /// Retrieve a file from the archive.
-    pub fn retrieve(
-        &self,
-        site: &Site,
-        model: Model,
-        init_time: chrono::NaiveDateTime,
-    ) -> Result<String, BufkitDataErr> {
-        let file_name: String = self.db_conn.query_row(
-            "SELECT file_name FROM files WHERE station_num = ?1 AND model = ?2 AND init_time = ?3",
-            &[
-                &site.station_num as &dyn rusqlite::types::ToSql,
-                &model.as_static_str() as &dyn rusqlite::types::ToSql,
-                &init_time as &dyn rusqlite::types::ToSql,
-            ],
-            |row| row.get(0),
-        )?;
-
-        let file = std::fs::File::open(self.data_root.join(file_name))?;
-        let mut decoder = flate2::read::GzDecoder::new(file);
-        let mut s = String::new();
-        decoder.read_to_string(&mut s)?;
-        Ok(s)
-    }
-
-    /// Retrieve the  most recent file
-    pub fn most_recent_file(&self, site: &Site, model: Model) -> Result<String, BufkitDataErr> {
-        let init_time = self.most_recent_init_time(site, model)?;
-        self.retrieve(site, model, init_time)
-    }
-
-    /// Retrieve all the soundings with data valid between the start and end times.
-    pub fn retrieve_all_valid_in(
-        &self,
-        start: chrono::NaiveDateTime,
-        end: chrono::NaiveDateTime,
-        site: &Site,
-        model: Model,
-    ) -> Result<Vec<String>, BufkitDataErr> {
-        let init_times = self.init_times_for_soundings_valid_between(start, end, site, model)?;
-
-        let string_data: Result<Vec<String>, _> = init_times
-            .into_iter()
-            .map(|init_t| self.retrieve(site, model, init_t))
-            .collect();
-
-        string_data
-    }
-
-    fn compressed_file_name(
-        &self,
-        site_id: &str,
-        model: Model,
-        init_time: chrono::NaiveDateTime,
-    ) -> String {
-        let file_string = init_time.format("%Y%m%d%HZ").to_string();
-
-        format!(
-            "{}_{}_{}.buf.gz",
-            file_string,
-            model.as_static_str(),
-            site_id.to_uppercase()
-        )
-    }
-
-    /// Get the file name this would have if uncompressed.
-    pub fn file_name(
-        &self,
-        site_id: &str,
-        model: Model,
-        init_time: &chrono::NaiveDateTime,
-    ) -> String {
-        let file_string = init_time.format("%Y%m%d%HZ").to_string();
-
-        format!(
-            "{}_{}_{}.buf",
-            file_string,
-            model.as_static_str(),
-            site_id.to_uppercase()
-        )
-    }
-
-    /// Remove a file from the archive.
-    pub fn remove(
-        &self,
-        site: &Site,
-        model: Model,
-        init_time: &chrono::NaiveDateTime,
-    ) -> Result<(), BufkitDataErr> {
-        let file_name: String = self.db_conn.query_row(
-            "SELECT file_name FROM files WHERE station_num = ?1 AND model = ?2 AND init_time = ?3",
-            &[
-                &site.station_num as &dyn rusqlite::types::ToSql,
-                &model.as_static_str() as &dyn rusqlite::types::ToSql,
-                init_time as &dyn rusqlite::types::ToSql,
-            ],
-            |row| row.get(0),
-        )?;
-
-        std::fs::remove_file(self.data_root.join(file_name)).map_err(BufkitDataErr::IO)?;
-
-        self.db_conn.execute(
-            "DELETE FROM files WHERE station_num = ?1 AND model = ?2 AND init_time = ?3",
-            &[
-                &site.station_num as &dyn rusqlite::types::ToSql,
-                &model.as_static_str() as &dyn rusqlite::types::ToSql,
-                init_time as &dyn rusqlite::types::ToSql,
-            ],
-        )?;
-
-        Ok(())
-    }
 }
 
 #[cfg(test)]
 mod unit {
     use super::*;
+    use crate::Model;
 
     use std::{convert::TryFrom, fs::read_dir};
 
@@ -705,7 +138,7 @@ mod unit {
         let test_data = get_test_data().expect("Error loading test data.");
 
         for (site, model, init_time, end_time, raw_data) in test_data {
-            arch.add(&site, model, init_time, end_time, &raw_data)?;
+            arch.add(site, model, init_time, end_time, &raw_data)?;
         }
         Ok(())
     }
@@ -882,7 +315,7 @@ mod unit {
             time_zone: Some(chrono::FixedOffset::west(7 * 3600)),
         };
 
-        arch.set_site_info(&zootown).expect("Error updating site.");
+        arch.update_site(&zootown).expect("Error updating site.");
 
         assert_eq!(arch.site_for_id("kmso").unwrap(), zootown);
         assert_ne!(arch.site_for_id("kmso").unwrap(), test_sites[2]);
@@ -935,7 +368,7 @@ mod unit {
     }
 
     #[test]
-    fn test_count() {
+    fn test_adding_duplicates() {
         let TestArchive {
             tmp: _tmp,
             mut arch,
@@ -943,27 +376,23 @@ mod unit {
 
         fill_test_archive(&mut arch).expect("Error filling test archive.");
 
-        // 7 and not 10 because of duplicate GFS models in the input.
-        assert_eq!(arch.count().expect("db error"), 7);
-    }
-
-    #[test]
-    fn test_count_init_times() {
-        let TestArchive {
-            tmp: _tmp,
-            mut arch,
-        } = create_test_archive().expect("Failed to create test archive.");
-
-        fill_test_archive(&mut arch).expect("Error filling test archive.");
+        let start = NaiveDate::from_ymd(2000, 1, 1).and_hms(0, 0, 0);
+        let end = NaiveDate::from_ymd(2100, 1, 1).and_hms(0, 0, 0);
 
         let kmso = arch.site_for_id("kmso").expect("Error retreiving MSO");
 
         assert_eq!(
-            arch.count_init_times(&kmso, Model::GFS).expect("db error"),
+            arch.init_times_for_soundings_valid_between(start, end, &kmso, Model::GFS)
+                .expect("db error")
+                .iter()
+                .count(),
             4
         );
         assert_eq!(
-            arch.count_init_times(&kmso, Model::NAM).expect("db error"),
+            arch.init_times_for_soundings_valid_between(start, end, &kmso, Model::NAM)
+                .expect("db error")
+                .iter()
+                .count(),
             3
         );
     }
@@ -976,12 +405,11 @@ mod unit {
         let test_data = get_test_data().expect("Error loading test data.");
 
         for (site, model, init_time, end_time, raw_data) in test_data {
-            arch.add(&site, model, init_time, end_time, &raw_data)
+            arch.add(site.clone(), model, init_time, end_time, &raw_data)
                 .expect("Failure to add.");
-            let site_obj = arch.site_for_id(&site.id.unwrap()).unwrap();
 
             let recovered_str = arch
-                .retrieve(&site_obj, model, init_time)
+                .retrieve(&site, model, init_time)
                 .expect("Failure to load.");
 
             assert!(raw_data == recovered_str);
